+++ apps/antalmanac/package.json
@@ -20,6 +20,7 @@
         "lint": "eslint src"
     },
     "dependencies": {
+        "@packages/antalmanac-types": "*",
         "@date-io/date-fns": "^2.16.0",
         "@emotion/react": "^11.10.5",
         "@emotion/styled": "^11.10.5",
@@ -48,6 +49,7 @@
         "moment": "^2.29.4",
         "moment-timezone": "^0.5.43",
         "notistack": "^2.0.8",
+        "peterportal-api-next-types": "1.0.0-alpha.6",
         "react": "^18.2.0",
         "react-big-calendar": "^1.6.4",
         "react-chartjs-2": "^5.2.0",
