+++ apps/antalmanac/src/lib/api/endpoints.ts
@@ -12,14 +12,6 @@ function endpointTransform(path: string) {
 
 export const LOOKUP_NOTIFICATIONS_ENDPOINT = endpointTransform('/api/notifications/lookupNotifications');
 export const REGISTER_NOTIFICATIONS_ENDPOINT = endpointTransform('/api/notifications/registerNotifications');
-export const RANDOM_AD_ENDPOINT = endpointTransform('/api/banners/getRandomAd');
-export const AD_IMAGE_ENDPOINT = endpointTransform('/api/banners/getAdImage');
-export const LOAD_DATA_ENDPOINT = endpointTransform('/api/users/loadUserData');
-export const LOAD_LEGACY_DATA_ENDPOINT = endpointTransform('/api/users/loadLegacyUserData');
-export const SAVE_DATA_ENDPOINT = endpointTransform('/api/users/saveUserData');
-export const ENROLLMENT_DATA_ENDPOINT = endpointTransform('/api/enrollmentData');
-export const NEWS_ENDPOINT = endpointTransform('/api/news');
-export const ZOTCOURSE_ENDPOINT = endpointTransform('/api/zotcourse/loadUserData');
 
 // PeterPortal API
 export const PETERPORTAL_GRAPHQL_ENDPOINT = 'https://api.peterportal.org/graphql';
