+++ apps/antalmanac/src/stores/AppStore.ts
@@ -1,7 +1,7 @@
 import { EventEmitter } from 'events';
 import { VariantType } from 'notistack';
 
-import { ScheduleCourse, ScheduleSaveState } from './schedule.types';
+import { ScheduleCourse, ScheduleSaveState } from '@packages/antalmanac-types';
 import { Schedules } from './Schedules';
 import { SnackbarPosition } from '$components/AppBar/NotificationSnackbar';
 import { CalendarEvent, CourseEvent } from '$components/Calendar/CourseCalendarEvent';
