+++ packages/types/package.json
@@ -6,7 +6,6 @@
   "types": "./src/index.ts",
   "type": "module",
   "scripts": {
-    "build": "tsc"
   },
   "dependencies": {
     "arktype": "1.0.14-alpha",
